Move runtime tests to own file
